requirements.txt
watchdog
pillow
netifaces
